+++ life.h
@@ -10,7 +10,7 @@ class Life{
         inline void set_a(int a_x,int a_y,int a_z);
         inline void cal_v();//计算v
         inline void cal_s();//计算位置
-        int x,y,z,ax,ay,az,vx,vy,vz;
+        double x,y,z,ax,ay,az,vx,vy,vz;
         int width,height;
     protected:
         int *id_pic;
